requirements.txt
streamlit
pandas
openpyxl
xlsxwriter
qrcode[pil]
altair
pillow
